fix: mudança no status e conteudo do retorno das funções de criar da API.

+++ backend/app/Services/JustificativaService.php
@@ -22,22 +22,50 @@ class JustificativaService
 
 	public function criarJustificativa(array $data)
 	{
-		return $this->justificativaRepository->create($data);
+		$justificativa = $this->justificativaRepository->firstOrNew($data);
+		if ($justificativa['id'] == null) {
+			$justificativa = $this->justificativaRepository->create($data);
+			return response()->json([
+				'message' => 'Justificativa criada com sucesso!',
+			], 201);
+		} else
+			throw new Exception('Justificativa já cadastrada');
 	}
 
 	public function procurarJustificativaPeloId($id)
 	{
-		return $this->justificativaRepository->find($id);
+		if ($id == null)
+			throw new Exception('ID não informado!');
+
+		$justificativa = $this->justificativaRepository->find($id);
+		if ($justificativa == null)
+			throw new Exception('Justificativa não encontrada!');
+
+		return $justificativa;
 	}
 
 	public function atualizarJustificativa(array $data, $id)
 	{
-		return $this->justificativaRepository->update($data, $id);
+		$justificativa = $this->justificativaRepository->find($id);
+		if ($justificativa == null)
+			throw new Exception('Justificativa não encontrada!');
+
+		$this->justificativaRepository->update($data, $id);
+		return response()->json([
+			'message' => 'Justificativa atualizada com sucesso!',
+		], 200);
 	}
 
 	public function excluirJustificativa($id)
 	{
-		return $this->justificativaRepository->delete($id);
+		$justificativa = $this->justificativaRepository->find($id);
+		if ($justificativa == null)
+			throw new Exception('Justificativa não encontrada!');
+
+		$this->justificativaRepository->delete($id);
+		return response()->json([
+			'message' => 'Justificativa excluída com sucesso!',
+		], 200);
 	}
 
 	public function retornaAJustificativaDoDiaDoFuncionario($funcionario, $data)
